add check if file exist after backup DIR_TO_NFS_WEEKLY

+++ data/cron/backup-universal.sh
@@ -1,11 +1,16 @@
 #!/bin/bash
 # env $DIR_TO_BACKUP $DIR_TO_NFS_WEEKLY $DIR_TO_NFS_MONTHLY $DAY_OF_WEEK (1..7) $DAY_OF_MONTH day of month (01..30)
+
 backup() {
   ALL_BACKUP="$(find $DIR_TO_BACKUP -type f -print0 | xargs -0)"
   for FILE in $ALL_BACKUP; do
     cp $FILE $DIR_TO_NFS_WEEKLY
-    echo "Weekly backup done: $FILE to NFS folder: $DIR_TO_NFS_WEEKLY"
-    rm $FILE
+    if [ -e $DIR_TO_NFS_WEEKLY/$(basename "$FILE") ]; then
+      rm $FILE
+      echo "Weekly backup done: to NFS folder: $DIR_TO_NFS_WEEKLY/$(basename "$FILE")"
+    else
+      echo "Some error, backup in $DIR_TO_NFS_MONTHLY not exists."
+    fi
   done
 }
 
@@ -13,7 +18,7 @@ rotate_backups() {
   LAST_BACKUP="$(ls -t $DIR_TO_NFS_WEEKLY | head -1)"
   ALL_BACKUP_WEEK="$(find $DIR_TO_NFS_WEEKLY -type f -print0 | xargs -0)"
   cp $DIR_TO_NFS_WEEKLY/$LAST_BACKUP $DIR_TO_NFS_MONTHLY
-  echo "Monthly backup done. last backup: $LAST_BACKUP"
+  echo "Monthly backup done. last backup: $DIR_TO_NFS_MONTHLY/$LAST_BACKUP"
   for FILE in $ALL_BACKUP_WEEK; do
     rm $FILE
   done
